Clean up for slurm out solution
Solution I found was to set up the 'output' option in the slurm_defaults 
cookiecutter secion of the slurm profile to the 
log_folder/{rule}/slurm.{rule}.{wildcards}.out. This has the jobID in it 
based on the profile's scripts so this, plus the new log (and eventually 
I'll add a stderr one), help to put all the outputs in an organzied way.