Updating bamlist and beagle rules
Updates to make sure all necessary files are created for each rule and are properly called in the command. However, testing is needed on threads and time for beagle creation. Just getting this version commited before Rackham goes down for maintenance so I can add more rules in the meantime.