Skip pruning if chunk is empty
As of now, the only way I can think of to avoid analyzing chunks that
are empty due to filtering is to use checkpoints.
I've had a lot of trouble with rerun triggers using checkpoints, so I
don't want to use them until they're more mature. Instead, the analyses
will be run, but with dummy files. Doesn't seem to be a big slow down.
Plus, completely empty chunks are an edge case I would think if the
chunk size is sufficient.