Redo sample depth and subsampling
Overhauls sample depth calculation to use ANGSD. ANGSD depth
calculations for unfiltered data match samtools depth closely in this
implementation. Post filtering depth is calculated in a way that it
is a good source for the subsampling to be based on in order to ensure
that subsampled depth is close to desired value after filtering.