Change GL calculating workflow
Before, GLs were calculated every time a saf or beagle file needed to
be made which was kind of computationally wasteful.

Now, they are calculated once for the whole dataset, then manipulated
be used as inputs for all combinations of samples. This should be a bit
more computationally efficient, at the expense of some space.