Allowing for choosing between Waples and NGSrelate
For relatedness, NGSrelate has much less overhead, and includes a form
of the methods in Waples et al. 2019. However, it makes the 2dsfs with
only variable sites, which may be a bit less accurate.

I've found them to get into the same ballpark, but users can choose
which, if any to run, or to run both.