Fixing bug in Saf calculation
There seems to be a bug in ANGSD that results in errors on sfs
generation if you build saf files from glfs or bcfs with >1 thread.
Multithreading seems to have no effect on Saf creation speed, so
limiting it to 1 seems to work.

It's hard to reproduce this error - it only happens randomly, so I only
see it when running datasets with many samples and sites. And really
only when I do 2dsfs for the Waples relatedness method.