Make clipbam fail eat less memory
Doing clipbam in a pipe with samtools sort leads to insane memory usage
with large genomes. Instead, will use shadows to hold some temp files.

Maybe down the line this can be split? Say make a rule for samtools
that can switch sorting based on wildcards on the fly and use it
throughout.