Update script based rules and their scripts
Before, I was using generalized scripts that were called in a shell
rule for these. This isn't great for ensuring reruns if the script is
modified or when running in an alternative working directory. These are
now no longer general, and can't be just used without the pipeline, but
are more robust within the pipeline framework. Old commits contain
the more generalized versions.