Calculating GLs on SAF and beagle generation
Using the text GLF worked, but since they are in beta, and didn't save
that much resources, I'm reverting to calculating the GLs at SAF and
beagle generation. This doesn't add up to too much more time and may be
less buggy. Keeping old code in, for potential return to GLF4 format.